Add hand animation and upload button
This commit adds a hand animation to indicate that the user can drag and drop files onto the page. It also adds an upload button for users who prefer to click to upload files, instead of the default file input button.

The height of the window was also adjusted to prevent Y scrolling on the page.